feat(currency): жолудь замість гривні — неподільна валюта
Гривня стає жолудем 1:1, і жолудь неподільний: копійок більше не існує
ніде — ні в базі, ні в жодному полі жодного клієнта.

Неподільність — властивість типу, а не перевірки, яку можна забути
викликати: усі дев'ять грошових колонок стали integer, тож дріб просто
нема де зберегти. RPC, що приймають суму від клієнта, кидають
FRACTIONAL_AMOUNT замість тихого округлення.

Нарахування перебудоване на точний цілочисельний накопичувач у
жолудь-секундах (accrued_acorn_seconds). Округлення порції втрачало б
залишок на кожному розрахунку — а розрахунок біжить і щодня по крону, і
на кожен запит виводу. Округлення накопиченого підсумку теж не годиться:
воно переоцінює всю історію за новою ставкою. Тому кожна нова порція
банкується за ставкою на той момент, у секундах×ставка — точно, цілими
числами, без жодного округлення. Живий хвіст тотожно дорівнює наступній
дельті, тож баланс не стрибає, коли спрацьовує крон.

Символ ₴ замінено іконкою жолудя: форматтер повертає лише число, а
іконку малює окремий компонент (AcornAmount.qml, KAcorns, .acn) — образ
не можна вклеїти всередину рядка так, щоб він сів на базову лінію.
Кольорова марка за замовчуванням: її двотонова шапочка тримає форму аж
до 12px, чого плаский силует не витримує. Тонування — лише на темних
заливках. У текстових контекстах (пуші, бот, нотатки) — слово зі
відмінюванням: 1 жолудь / 2 жолуді / 5 жолудів.

Міграція боєвих даних зберігає інваріант «баланс = сума ledger»:
історію не переписано заднім числом, розбіжність від округлення несе
одна видима корекція на дитину, а контрольні значення лишаються в
acorn_migration_audit.

Клієнти декодують гроші поблажливо (AcornSerializer) і показують помилку
завантаження замість нуля: строгий Int відкидав дробове значення, а
kotlinx перериває розбір усієї відповіді — через це баланс, історія і
виплати зникали разом, і екран показував упевнений 0.

+++ android/app/src/main/kotlin/com/kabanchiki/app/core/data/BalanceRepository.kt
@@ -53,7 +53,7 @@ class BalanceRepository @Inject constructor(
     }
 
     /** amount == null cashes out the whole balance. Returns the new payout id. */
-    suspend fun requestWithdrawal(amount: Double? = null): String {
+    suspend fun requestWithdrawal(amount: Int? = null): String {
         val result = client.postgrest.rpc(
             "request_withdrawal",
             buildJsonObject { if (amount != null) put("p_amount", amount) },
